Simplify error handling
When we're bubbling up a Python-space exception as part of `PyErr`, we
don't need to reproduce its message, because it already will be
displayed/inspectable.  In cases where we don't have any meaningful
additional context to add, we could either refactor `ParseError` to be
an `enum` with an explicit raw `Py` variant, or we can (as this commit
does) simply say we've caught an exception and chained it on.
