+++ crates/qasm2/src/bytecode.rs
@@ -10,22 +10,8 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-use crate::error::ParseError;
 use num_bigint::BigUint;
 use pyo3::prelude::*;
-pyo3::import_exception!(qiskit.qasm2.exceptions, QASM2ParseError);
-
-/// Convert a `ParseError` from the pyo3-free parsing modules into the `QASM2ParseError`
-/// Python exception, at the boundary where results cross back into Python space.
-impl From<ParseError> for PyErr {
-    fn from(e: ParseError) -> PyErr {
-        let py_err = QASM2ParseError::new_err(e.message);
-        if let Some(source) = e.source {
-            Python::attach(|py| py_err.set_cause(py, Some(*source)));
-        }
-        py_err
-    }
-}
 
 use crate::expr::Expr;
 use crate::ext::ClassicalCallableExt;
