+++ crates/qasm2/src/ext.rs
@@ -125,45 +125,31 @@ pub struct CustomClassical {
 impl CustomClassical {
     #[new]
     #[pyo3(text_signature = "(name, num_params, callable, /)")]
-    fn __new__(name: String, num_params: usize, callable: Py<PyAny>) -> Self {
+    fn __new__(name: String, num_params: usize, callable: Py<PyAny>) -> PyResult<Self> {
+        let lift_pyerr = |err| {
+            ParseError::with_source(
+                "caught exception when constant folding with user-defined function",
+                err,
+            )
+        };
         // Capture Py<PyAny> inside the Arc<dyn Fn> closure so ClassicalCallableExt itself
         // has no pyo3 type in its definition.
         let f: ClassicalFn = Arc::new(move |params: &[f64]| {
             Python::attach(|py| {
-                // f64::IntoPyObject::Error = Infallible, so this cannot fail.
-                let py_args = pyo3::types::PyTuple::new(py, params)
-                    .expect("f64 -> PyFloat conversion is infallible");
-                let result = callable.call1(py, py_args).map_err(|e| {
-                    let desc = e
-                        .value(py)
-                        .str()
-                        .map(|s| s.to_string())
-                        .unwrap_or_else(|_| {
-                            let type_name = e
-                                .get_type(py)
-                                .qualname()
-                                .map(|n| n.to_string())
-                                .unwrap_or_else(|_| "?".to_string());
-                            format!("<{type_name}: unprintable>")
-                        });
-                    // Keep the original PyErr as `source` so that the traceback is preserved.
-                    ParseError::with_source(
-                        format!("caught exception when constant folding: {desc}"),
-                        e,
-                    )
-                })?;
+                let py_args = pyo3::types::PyTuple::new(py, params).map_err(lift_pyerr)?;
+                let result = callable.call1(py, py_args).map_err(lift_pyerr)?;
                 result.extract::<f64>(py).map_err(|e| {
                     ParseError::with_source(
-                        "user-provided classical function returned non-float".to_owned(),
+                        "user-defined function returned non-float during constant folding",
                         e,
                     )
                 })
             })
         });
-        Self {
+        Ok(Self {
             name,
             callable: ClassicalCallableExt::Custom { num_params, f },
-        }
+        })
     }
 
     /// Get a list of all the custom classical instructions that are built into Qiskit, but not part
